Ensure that variable bindings can live across calls to Lox::run()
There were two problems:

1. We were attempting to store binding keys as &str, pointing to the
location in the source code where they are defined, but the source
code buffer does not live across calls to Lox::run(). We now store
binding keys as String.

2. The environment in which we store bindings was being created inside
Lox::run(), which meant that it too went out of scope between run()
calls. We now use an Environment trait that Lox itself implements
allowing it to be passed as the execution environment.