Refactor interpreter to use an explicit stack
Up to now we have used the Rust stack as the stack of the interpreter
when evaluating code. This becomes a problem when introducing closures
as they execute in the context in which they were defined (their
captured environment), which is different to that in which they are
called. This means that we need to be able to change the scope of
execution independently from the Rust stack.

We now use an explicit stack which has frames pushed and popped
whenever a new scope is entered or exited. This should allow us to
push the captured environment of a closure before executing it thereby
ensuring that it executes with the correct scope.