Correctly resolve closure variables
When an intermediate environment is created with a variable of the
same name as one referenced by the closure, the new variable shadows
the one in the closure's environment leading to incorrect resolution.

By using a second resolution pass we now statically resolve all
variables to the scope in which they live, which solves this problem.