Fix recursion handling
The arguments to functions were being evaluated after the new function
scope was created, meaning that any referenced variables were at the
wrong depth to be found. This prevented recursive functions from
working.

The fix is to evaluate all of the arguments in the current scope
before pushing the new scope, then define the results in the new scope
once it is pushed.